requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
chromadb==0.3.21
click==8.1.7
clickhouse-connect==0.6.14
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.5.14
duckdb==0.9.0
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gradio==3.27.0
gradio_client==0.6.0
greenlet==3.0.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.148
linkify-it-py==2.0.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
openai==0.27.4
openapi-schema-pydantic==1.2.4
orjson==3.9.7
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
posthog==3.0.2
pydantic==1.10.13
pydub==0.25.1
PyMuPDF==1.22.2
pyparsing==3.1.1
pypdf==3.8.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rpds-py==0.10.4
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.3.3
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.6
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
zstandard==0.21.0
